# docs/openapi/teamclaw-api.v1.yaml
openapi: 3.1.0
info:
  title: TeamClaw Cloud API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.teamclaw.local
security:
  - bearerAuth: []
...
paths:
...
  /v1/teams/{teamId}/workspace-config:
    get:
      operationId: getTeamWorkspaceConfig
      summary: Get merged team workspace config (share mode + litellm)
      description: |
        Returns the merged team workspace config introduced in the team-share
        onboarding refactor: share mode, git remote/auth (when applicable),
        sync mode, and the provisioned LiteLLM team id. The legacy default
        workspace shape now lives at `GET /v1/teams/{teamId}/workspace-defaults`.
      tags: [Teams]
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Merged team workspace config.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MergedWorkspaceConfig"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    put:
      operationId: putTeamWorkspaceConfig
      summary: Put team workspace config (legacy defaults shape)
      description: |
        Upserts the legacy team workspace defaults shape
        (`defaultWorkspaceId`, `pinnedWorkspaceIds`). Does not modify
        share-mode or litellm fields — those are managed via the
        `/share-mode` and `/litellm/setup` endpoints.
      tags: [Teams]
      parameters:
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamWorkspaceConfig"
      responses:
        "200":
          description: Upserted team workspace config.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWorkspaceConfig"
        "400":
          $ref: "#/components/responses/Error"
...
  /v1/teams/{teamId}/share-mode:
    post:
      operationId: enableTeamShareMode
      summary: Enable team share mode (one-shot lock)
      description: |
        Atomically locks the team's share mode to one of `oss`, `managed_git`,
        or `custom_git`. Once locked, the mode cannot be changed — a second
        call returns 409 `share_mode_locked`. For `managed_git` and
        `custom_git`, a `gitConfig.remoteUrl` is required. For `custom_git`,
        the caller must also supply `gitConfig.authKind` and
        `gitConfig.credentialRef`.
      tags: [Teams]
      parameters:
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableShareModeRequest"
      responses:
        "200":
          description: Team after share mode is enabled.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
    get:
      operationId: getTeamShareMode
      summary: Get team share mode status
      tags: [Teams]
      parameters:
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Current team share status. `mode` is null when share has not been opened.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareStatus"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  headers:
    RequestId:
      schema:
        type: string
  parameters:
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Error:
      description: Error response.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - missing_auth
                - invalid_json
                - validation_failed
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - upstream_unavailable
                - internal
            message:
              type: string
            requestId:
              type: string
    Team:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type:
            - string
            - "null"
        createdAt:
          type:
            - string
            - "null"
...
    TeamWorkspaceConfig:
      type: object
      required: [teamId, defaultWorkspaceId]
      properties:
        teamId:
          type: string
        defaultWorkspaceId:
          type:
            - string
            - "null"
        pinnedWorkspaceIds:
          type: array
          items:
            type: string
        updatedAt:
          type:
            - string
            - "null"
    TeamShareMode:
      type: string
      enum: [oss, managed_git, custom_git]
      description: |
        Team share mode. `oss` uses Alibaba OSS / WebDAV; `managed_git` uses a
        TeamClaw-provisioned git remote; `custom_git` uses a caller-supplied
        git remote with explicit credential metadata.
    EnableShareModeRequest:
      type: object
      required: [mode]
      properties:
        mode:
          $ref: "#/components/schemas/TeamShareMode"
        gitConfig:
          type: object
          description: |
            Required when `mode` is `managed_git` or `custom_git`. For
            `custom_git`, `authKind` and `credentialRef` are also required.
            `credentialRef` references a device-local secret entry; the FC
            never sees the raw credential material.
          required: [remoteUrl]
          properties:
            remoteUrl:
              type: string
            authKind:
              type: string
              enum: [ssh_key, https_token]
            credentialRef:
              type: string
    ShareStatus:
      type: object
      required: [mode, enabledAt, gitRemoteUrl, gitAuthKind]
      properties:
        mode:
          oneOf:
            - $ref: "#/components/schemas/TeamShareMode"
            - type: "null"
        enabledAt:
          type:
            - string
            - "null"
          format: date-time
        gitRemoteUrl:
          type:
            - string
            - "null"
        gitAuthKind:
          type:
            - string
            - "null"
          enum: [ssh_key, https_token, null]
    MergedWorkspaceConfig:
      type: object
      required: [shareMode, gitRemoteUrl, gitAuthKind, syncMode, litellmTeamId]
      properties:
        shareMode:
          oneOf:
            - $ref: "#/components/schemas/TeamShareMode"
            - type: "null"
        gitRemoteUrl:
          type:
            - string
            - "null"
        gitAuthKind:
          type:
            - string
            - "null"
          enum: [ssh_key, https_token, null]
        syncMode:
          type:
            - string
            - "null"
        litellmTeamId:
          type:
            - string
            - "null"
...